requirement.txt
python-dotenv
fastmcp
pydantic
yt-dlp
assemblyai
